dm-xnat-upload handles new XNAT project
Previously, dm-xnat-upload looked for the arc001 and if it wasn't found, would exit.
This is problematic because when an XNAT project is first created it doesn't have an arc001
folder, which meant you had to manually upload a zip before running dm-xnat-upload.
But, no more!

I also added a check that the correct number of arguments were given to this script.