Preserving file mtime on xnat export
ctime is updated when permissions/ownership change, whereas mtime should only
change when the contents of a file change.

dm-qc-todo was reporting that qc documents were older than the source files
they depended on and this turned out to be for two reasons:

 1. dm-qc-todo was comparing file ctimes but ctimes can change (and do because
 of the nightly permission/ownership reset) without it indicating a file
 change.

 2. When I fixed dm-qc-todo to use mtime, it reported several qc documents as
 being older than the spiral scan files located in the RESOURCES folder. This
 happened because xnat-extract uses rsync to sync the RESOURCES folder but
 wasn't preserving mtimes, it was reseting them to the current time.

With this commit, mtimes are preserved on exported RESOURCES
like spiral scans, and dm-qc-todo more sensibly compares
mtime file timestamps.