Added setup.py and conda environment spec
This makes it possible to install datman. I haven't tested it
with all of the various pipelines so there may be bits and pieces
missing.

Also, I have defined the dependencies in setup.py and environment.yml.
There may be a better way of doing this so that we just have one file.