Bugfix: Resource files being uploaded every night.
By resource files were being added to a 'default' subfolder on xnat. By default this folder was given the name 'No Label'.
If another folder e.g. 'BEHAV' existed resource files were being added to this folder.
We were only checking a single subfolder for the presence of resource files, causing them to be uploaded again if this folder wasn't the one we were looking at.

Modified scripts so new resource files will be added to a named subfolder 'MISC'.
Also modified so all subfolders are checked for presence of resource files.