Add dm_parse_ea.py
The very messy version of dm_parse_ea that was used to parse EA task log files in early October 2019. This will be eventually updated to not be hardcoded and to actually use datman. but for now just committing this to datman from my https://github.com/gabiherman/parse_ea_task so it can be found centrally. It would probably be better suited for rigorous code review a little later down the line...