[FIX] Assume scans without image_type are not derived (if they have a tag)
TAY01 ASL scans weren't being pulled down because they had no image_type
listed and were being assumed to be derived. This swaps the order
of the checks so that by the time we check if a scan is derived we
already know it has a matching tag and therefore it's very likely
a scan we want to download.